// optoweb/src/lib/evaluacion-types.ts
export type SaveEvaluacionAction = (
  prevState: { error?: string } | null,
  formData: FormData
) => Promise<{ error?: string } | null>;

export type EvaluacionDraft = {
  fecha: string;
  motivoConsulta: string;
  sintomasSignos: string;
  antecedentesPersonalesOculares: string;
  antecedentesPersonalesSistemicos: string;
  antecedentesFamiliaresOculares: string;
  antecedentesFamiliaresSistemicos: string;
  medicacion: string;
  alergias: string;
  necesidadVisual: string;
  avScAo: string;
  avScOdLejos: string;
  avScOiLejos: string;
  avCcAoPx: string;
  avCcOdLejos: string;
  avCcOiLejos: string;
  estereopsisValor: string;
  estereopsisSegundos: string;
  lang: string;
  worth: string;
  ishihara: string;
  farnsworth: string;
  schirmerOd: string;
  schirmerOi: string;
  osdiPuntuacion: string;
  osdiClasificacion: string;
  sensibilidadContraste: string;
  sensibilidadFrecuencia: string;
  amsler: string;
  campoVisual: string;
  campoVisualDescripcion: string;
  phOd: string;
  phOi: string;
  kappaOd: string;
  kappaOi: string;
  hirshberg: string;
  duccionesOd: string;
  duccionesOi: string;
  versionesAo: string;
  coverTest6m: string;
  coverTest40cm: string;
  coverTest10cm: string;
  ppcOr: string;
  ppcLuz: string;
  ppcFrl: string;
  reflejoFotomotor: string;
  reflejoConsensual: string;
  reflejoAcomodativo: string;
  objOdEsf: string;
  objOdCil: string;
  objOdEje: string;
  objOiEsf: string;
  objOiCil: string;
  objOiEje: string;
  subjOdEsf: string;
  subjOdCil: string;
  subjOdEje: string;
  subjOiEsf: string;
  subjOiCil: string;
  subjOiEje: string;
  recetaOdEsf: string;
  recetaOdCil: string;
  recetaOdEje: string;
  recetaOdAv: string;
  recetaOiEsf: string;
  recetaOiCil: string;
  recetaOiEje: string;
  recetaOiAv: string;
  addCercaOd: string;
  addCercaOi: string;
  addAv: string;
  isAddAo: string;
  dipLejos: string;
  dipCerca: string;
  prismaOdValor: string;
  prismaOdBase: string;
  prismaOiValor: string;
  prismaOiBase: string;
  k1Od: string;
  k2Od: string;
  k1Oi: string;
  k2Oi: string;
  lcOdEsf: string;
  lcOdCil: string;
  lcOdEje: string;
  lcOiEsf: string;
  lcOiCil: string;
  lcOiEje: string;
  lcRadioBaseOd: string;
  lcRadioBaseOi: string;
  lcOdDia: string;
  lcOiDia: string;
  lcLaboratorio: string;
  lcTipoLente: string;
  lcMaterial: string;
  lcFechaAdaptacion: string;
  lcObservaciones: string;
  contactologia: string;
  diagnosticoOd: string;
  diagnosticoOi: string;
  balanceOd: string;
  balanceOi: string;
  otrosPresbicia: string;
  otrosAnisometropia: string;
  otrosAmbliopia: string;
  autoPresbicia: string;
  autoAnisometropia: string;
  autoAmbliopia: string;
  planTratamiento: string;
  citaEstado: string;
  diagnosticoResumen: string;
  indicaciones: string;
  proximaCita: string;
  notasFinales: string;
};

export type DraftUpdateFn = <K extends keyof EvaluacionDraft>(
  key: K,
  value: EvaluacionDraft[K]
) => void;